svm: move StringUtil to shared module
