+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/util/CPUTypeRISCV64.java
@@ -42,8 +42,8 @@ import org.graalvm.nativeimage.Platforms;
 import com.oracle.svm.core.option.SubstrateOptionsParser;
 import com.oracle.svm.core.util.UserError;
 import com.oracle.svm.hosted.NativeImageOptions;
+import com.oracle.svm.shared.util.StringUtil;
 import com.oracle.svm.util.GuestAccess;
-import com.oracle.svm.util.StringUtil;
 
 import jdk.vm.ci.riscv64.RISCV64;
 import jdk.vm.ci.riscv64.RISCV64.CPUFeature;
