+++ substratevm/src/com.oracle.svm.util/src/com/oracle/svm/util/GlobUtils.java
@@ -30,6 +30,8 @@ import java.util.List;
 import java.util.function.Consumer;
 import java.util.regex.Pattern;
 
+import com.oracle.svm.shared.util.StringUtil;
+
 public final class GlobUtils {
 
     private GlobUtils() {
